// src/app/models/bond.ts
export class Bond {
    id: string;
    date: any;
    name: string;
    number: string;
    onlineid: string;
    passwordhint: string;
    memorableplace: string;
    memorablename: string;
    memorabledate: string;
    initialcash: number;
    balance: number;
    interestrate: number;
    interestpayable: number;
    frequency: string;
    maturity: string;
    annualinterest: number;
    taxable: string;
    notes: string;
    forecast: boolean;
}